requirements.txt
matplotlib==3.2.1
numpy==1.18.3
opencv_python==4.4.0.40
Pillow==8.4.0
torch==1.10.0
torchvision==0.11.2
